refactor plugins file
