+++ lua/plugin_configs/alpha.lua
@@ -1,0 +1,23 @@
+local M = {}
+local logo = [[
+─────────────────────────────────────────────────────────────────────────────────────────────────────────────────
+─────────██████─██████████████─██████████████─██████──██████────────████████████───██████████████─██████──██████─
+─────────██░░██─██░░░░░░░░░░██─██░░░░░░░░░░██─██░░██──██░░██────────██░░░░░░░░████─██░░░░░░░░░░██─██░░██──██░░██─
+─────────██░░██─██░░██████░░██─██░░██████████─██░░██──██░░██────────██░░████░░░░██─██░░██████████─██░░██──██░░██─
+─────────██░░██─██░░██──██░░██─██░░██─────────██░░██──██░░██────────██░░██──██░░██─██░░██─────────██░░██──██░░██─
+─────────██░░██─██░░██──██░░██─██░░██████████─██░░██████░░██────────██░░██──██░░██─██░░██████████─██░░██──██░░██─
+─────────██░░██─██░░██──██░░██─██░░░░░░░░░░██─██░░░░░░░░░░██────────██░░██──██░░██─██░░░░░░░░░░██─██░░██──██░░██─
+─██████──██░░██─██░░██──██░░██─██████████░░██─██░░██████░░██────────██░░██──██░░██─██░░██████████─██░░██──██░░██─
+─██░░██──██░░██─██░░██──██░░██─────────██░░██─██░░██──██░░██────────██░░██──██░░██─██░░██─────────██░░░░██░░░░██─
+─██░░██████░░██─██░░██████░░██─██████████░░██─██░░██──██░░██─██████─██░░████░░░░██─██░░██████████─████░░░░░░████─
+─██░░░░░░░░░░██─██░░░░░░░░░░██─██░░░░░░░░░░██─██░░██──██░░██─██░░██─██░░░░░░░░████─██░░░░░░░░░░██───████░░████───
+─██████████████─██████████████─██████████████─██████──██████─██████─████████████───██████████████─────██████─────
+─────────────────────────────────────────────────────────────────────────────────────────────────────────────────
+]]
+
+M.otps = function()
+  local dashboard = require("alpha.themes.dashboard")
+  dashboard.section.header.val = vim.split(logo, "\n")
+end
+
+return M
