+++ lua/plugin_configs/indent_blankline.lua
@@ -1,0 +1,11 @@
+local M = {}
+
+M.opts = function()
+  return {
+    show_current_context_start = true,
+    show_trailing_blankline_indent = true,
+    show_current_context = true,
+  }
+end
+
+return M
