+++ lua/plugin_configs/treesitter.lua
@@ -1,0 +1,7 @@
+local M = {}
+
+M.opts = function(_, opts)
+  table.insert(opts.ensure_installed, "jsdoc")
+end
+
+return M
